+++ app/src/lib/build-info.js
@@ -1,2 +1,2 @@
-export const APP_BUILD = 'v1.7.1'
-export const BUILD_DATE = '2026-06-08 15:30'
+export const APP_BUILD = 'v1.7.2'
+export const BUILD_DATE = '2026-06-08 16:30'